# internal/tmpl: overhaul Tree.WriteTo for clarity