@josharian 2023-10-20
The easiest way to review is probably to start by reading the core of the new implementation: https://github.com/canopyclimate/golive/blob/e017dadd65bfe602e8ff3b07671d336792eab64e/internal/tmpl/tree.go#L229-L291

and then spidering out from there and/or looking at the diff.

Happy to answer questions or help in any way.